Refactor structure for cleaner diff
