+++ frontend/src/previous_card.tsx
@@ -2,15 +2,15 @@ import { Timestamp } from "@bufbuild/protobuf";
 import { Score } from "./gen/rpc/wordvault/api_pb";
 import React, { useState } from "react";
 import { Alert, Button, Group, Popover, Stack, Text } from "@mantine/core";
-import { CardStats } from "./card_stats";
-import { CardStat, ParsedFsrsCard } from "./types";
+import { CardStats } from "./stats/card_recall_stats";
+import { CardRecallStat, ParsedFsrsCardStats } from "./stats/types";
 
 export interface HistoryEntry {
   score: Score;
   alphagram: string;
   nextScheduled: Timestamp;
-  cardRepr: ParsedFsrsCard;
-  previousCardRepr: ParsedFsrsCard;
+  cardRepr: ParsedFsrsCardStats;
+  previousCardRepr: ParsedFsrsCardStats;
 }
 
 interface PreviousCardProps {
@@ -41,7 +41,7 @@ const PreviousCard: React.FC<PreviousCardProps> = ({
               size: "sm",
               c: "dimmed",
             }}
-            excludeStats={new Set([CardStat.LAST_SEEN])}
+            excludeStats={new Set([CardRecallStat.LAST_SEEN])}
           />
           <Text size="sm" c="dimmed">
             Score:{" "}
